# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Student OS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /courses:
    get:
      operationId: listCourses
      tags: [courses]
      responses:
        "200":
          description: List of courses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Course"
    post:
      operationId: createCourse
      tags: [courses]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
  /courses/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    patch:
      operationId: updateCourse
      tags: [courses]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
        "404":
          description: Not found
    delete:
      operationId: deleteCourse
      tags: [courses]
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ---------- Timetable ----------
...
  /attendance-records:
    get:
      operationId: listAttendanceRecords
      tags: [attendance]
      parameters:
        - name: courseId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of attendance records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceRecord"
    post:
      operationId: createAttendanceRecord
      tags: [attendance]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceRecordInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRecord"
  /attendance-records/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    patch:
      operationId: updateAttendanceRecord
      tags: [attendance]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceRecordUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRecord"
        "404":
          description: Not found
    delete:
      operationId: deleteAttendanceRecord
      tags: [attendance]
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
...
  /grade-entries:
    get:
      operationId: listGradeEntries
      tags: [grades]
      responses:
        "200":
          description: List of grade entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GradeEntry"
    post:
      operationId: createGradeEntry
      tags: [grades]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GradeEntryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GradeEntry"
  /grade-entries/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    patch:
      operationId: updateGradeEntry
      tags: [grades]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GradeEntryUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GradeEntry"
        "404":
          description: Not found
    delete:
      operationId: deleteGradeEntry
      tags: [grades]
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
...
  /notes/{id}/summarize:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    post:
      operationId: summarizeNote
      tags: [ai]
      summary: Generate an AI summary for a note and save it
      responses:
        "200":
          description: Note with generated summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        "404":
          description: Not found

  # ---------- Focus Shield ----------
  /blocklist-items:
    get:
      operationId: listBlocklistItems
      tags: [focus-shield]
      summary: List the distraction apps/sites the user wants to avoid
      responses:
        "200":
          description: List of blocklist items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BlocklistItem"
    post:
      operationId: createBlocklistItem
      tags: [focus-shield]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlocklistItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlocklistItem"
  /blocklist-items/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    delete:
      operationId: deleteBlocklistItem
      tags: [focus-shield]
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
...
components:
  schemas:
...
    Course:
      type: object
      required: [id, name, credits, color, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        code: { type: ["string", "null"] }
        instructor: { type: ["string", "null"] }
        color: { type: string }
        credits: { type: number }
        createdAt: { type: string, format: date-time }
    CourseInput:
      type: object
      required: [name, credits]
      properties:
        name: { type: string, minLength: 1 }
        code: { type: string }
        instructor: { type: string }
        color: { type: string }
        credits: { type: number }
    CourseUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        code: { type: string }
        instructor: { type: string }
        color: { type: string }
        credits: { type: number }
...
    AttendanceRecord:
      type: object
      required: [id, courseId, date, status]
      properties:
        id: { type: integer }
        courseId: { type: integer }
        date: { type: string, format: date }
        status: { type: string, enum: [present, absent, excused] }
    AttendanceRecordInput:
      type: object
      required: [courseId, date, status]
      properties:
        courseId: { type: integer }
        date: { type: string, format: date }
        status: { type: string, enum: [present, absent, excused] }
    AttendanceRecordUpdate:
      type: object
      properties:
        courseId: { type: integer }
        date: { type: string, format: date }
        status: { type: string, enum: [present, absent, excused] }
...
    Note:
      type: object
      required: [id, title, content, tags, createdAt, updatedAt]
      properties:
        id: { type: integer }
        title: { type: string }
        content: { type: string }
        courseId: { type: ["integer", "null"] }
        summary: { type: ["string", "null"] }
        tags:
          type: array
          items: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    GradeEntry:
      type: object
      required: [id, semesterName, courseName, credits, gradePoint]
      properties:
        id: { type: integer }
        semesterName: { type: string }
        courseName: { type: string }
        credits: { type: number }
        gradePoint: { type: number }
    GradeEntryInput:
      type: object
      required: [semesterName, courseName, credits, gradePoint]
      properties:
        semesterName: { type: string, minLength: 1 }
        courseName: { type: string, minLength: 1 }
        credits: { type: number }
        gradePoint: { type: number }
    GradeEntryUpdate:
      type: object
      properties:
        semesterName: { type: string, minLength: 1 }
        courseName: { type: string, minLength: 1 }
        credits: { type: number }
        gradePoint: { type: number }
...
    BlocklistItem:
      type: object
      required: [id, label, category, createdAt]
      properties:
        id: { type: integer }
        label: { type: string }
        category:
          type: string
          enum: [social, video, games, messaging, other]
        createdAt: { type: string, format: date-time }
    BlocklistItemInput:
      type: object
      required: [label]
      properties:
        label: { type: string, minLength: 1 }
        category:
          type: string
          enum: [social, video, games, messaging, other]
...