# lib/api-zod/src/generated/types/blocklistItem.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Student OS API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BlocklistItemCategory } from './blocklistItemCategory';

export interface BlocklistItem {
  id: number;
  label: string;
  category: BlocklistItemCategory;
  createdAt: Date;
}